+++ src/routes/edit.tsx
@@ -1,9 +1,12 @@
 import { createFileRoute } from "@tanstack/react-router";
 import { useEffect, useState } from "react";
 
+import type { LoadingData, PhotoBody } from "@/types";
+
+import { DEFAULT_WINDOW_TITLE } from "@/constants";
+
 import ImageEditor from "@/frontend/components/ImageEditor";
 import LoadingOverlay from "@/frontend/components/LoadingOverlay";
-import type { LoadingData, PhotoBody } from "@/types";
 
 const fetchLocalFile = async (data: PhotoBody) => {
   const response = await fetch(`file://${data.directory}/${data?.edited || data.name}`);
@@ -12,17 +15,19 @@ const fetchLocalFile = async (data: PhotoBody) => {
 };
 
 const EditPage = () => {
+  const [query, setQuery] = useState<string>(
+    new URLSearchParams(window.location.search).get("data")!,
+  );
   const [loading, setLoading] = useState<LoadingData>({ show: true });
   const [data, setData] = useState<PhotoBody | null>(null);
   const [file, setFile] = useState<File | null>(null);
 
-  // For hash routing
-  const queryValue = new URLSearchParams(window.location.search).get("data")!;
-
   useEffect(() => {
     async function fetchData() {
-      const parsedData = JSON.parse(atob(queryValue)) as PhotoBody;
+      const parsedData = JSON.parse(atob(query)) as PhotoBody;
       setData(parsedData);
+
+      document.title = `${DEFAULT_WINDOW_TITLE} - ${parsedData.directory}/${parsedData.name}`;
 
       const response = await fetchLocalFile(parsedData);
       setFile(response);
@@ -31,12 +36,12 @@ const EditPage = () => {
     }
 
     fetchData();
-  }, [queryValue]);
+  }, [query]);
 
   return (
     <>
       <LoadingOverlay data={loading} />
-      {data && file && <ImageEditor data={data} image={file} />}
+      {data && file && <ImageEditor data={data} image={file} setQueryCallback={setQuery} />}
     </>
   );
 };